Flow.Launcher.Plugin.Translate/requirements.txt
pyperclip
libretranslatepy
flox-lib==0.19.6